Add files via upload
This is a C++ code of parallel scan using MPI. In MPI, there is already a scan function(exscan and inscan), but this code is what algorithm is implemented in those function. It is quite fascinating! The pseudo code that I referred to is here http://www2.hawaii.edu/~nodari/teaching/f17/slides/L06-segmented-prefix-sums.pdf